Create 1_Numpy_Project_CrunchieMunchies.py
CrunchieMunchies
You work in marketing for a food company YummyCorps, which is developing a new kind of tasty, wholesome cereal called CrunchieMunchies. You want to demonstrate to consumers how healthy your cereal is in comparison to other leading brands, so you’ve dug up nutritional data on several different competitors.
Look over the cereal.csv file. This file contains the reported calorie amounts for different cereal brands. Load the data from the file.
There are 60 calories per serving of CrunchieMunchies. How much higher is the average calorie count of your competition?
Does the average calorie count adequately reflect the distribution of the dataset?
Calculate median and compare it to the average.
While the median demonstrates that at least half of our values are over 100 calories, it would be more impressive to show that a significant portion of the competition has a higher calorie count that CrunchieMunchies.
Let’s calculate the percentage of cereals that have more than 60 calories per serving.
Wow! That’s a really high percentage. That’s going to be very useful when we promote CrunchieMunchies. But one question is, how much variation exists in the dataset? Can we make the generalization that most cereals have around 100 calories or is the spread even greater?